# docs/openapi-spec/public/svc-web-analyzer-swagger-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Web Page Analyzer API",
    "description": "A web application that analyzes web pages and provides detailed information about:\n- HTML version\n- Page title\n- Heading counts by level\n- Internal and external links\n- Inaccessible links\n- Login form detection\n\n## API Versioning\n\nThis API uses semantic versioning and supports multiple versioning strategies:\n\n### Version Strategy\n- **URL Path Versioning**: `/v1/` (primary method)\n- **Header Versioning**: `API-Version: v1` header (alternative)\n- **Content Type Versioning**: `application/vnd.web-analyzer.v1+json` (for specific operations)\n\n### Version Information\n- All responses include `API-Version` header indicating the version used\n- Version-specific changes are documented in the changelog\n- Breaking changes require major version increment\n\n## Security\n\nThis API uses PASETO token authentication:\n- **PASETO tokens**: Platform Authentication Security Token Exchange and Operations - enhanced security tokens with issuer validation\n\n## Security Headers\n\nAll responses include standard security headers:\n- `X-Content-Type-Options: nosniff`\n- `X-Frame-Options: DENY`\n- `X-XSS-Protection: 1; mode=block`\n- `Strict-Transport-Security: max-age=31536000; includeSubDomains`\n- `Content-Security-Policy: default-src 'self'`\n- `Referrer-Policy: strict-origin-when-cross-origin`\n- `Permissions-Policy: camera=(), microphone=(), geolocation=()`\n",
    "version": "1.0.0",
...
  "servers": [
...
  "security": [
    {
      "PasetoAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/analysis/{analysisId}/events": {
      "get": {
        "summary": "Get real-time analysis progress",
        "description": "Server-Sent Events endpoint for real-time analysis progress updates.\nThis endpoint streams live updates about the analysis progress.\n",
        "operationId": "getAnalysisEvents",
        "tags": [
          "Real-time"
        ],
        "security": [
          {
            "PasetoAuth": []
          }
        ],
        "parameters": [
          {
            "name": "API-Version",
            "in": "header",
            "required": false,
            "description": "API version to use for this request. If not specified, defaults to v1.\nSupported versions: v1\n",
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ],
              "default": "v1"
            },
            "example": "v1"
          },
          {
            "name": "analysisId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The unique identifier of the analysis"
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of analysis progress",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "progress_events": {
                    "summary": "SSE progress events",
                    "value": "event: started\ndata: {\"analysis_id\": \"550e8400-e29b-41d4-a716-446655440000\", \"status\": \"started\", \"timestamp\": \"2025-01-15T10:30:00Z\"}\n\nevent: progress\ndata: {\"step\": \"fetching_page\", \"progress\": 25, \"message\": \"Fetching page content...\", \"timestamp\": \"2025-01-15T10:30:05Z\"}\n\nevent: progress\ndata: {\"step\": \"parsing_html\", \"progress\": 50, \"message\": \"Parsing HTML content...\", \"timestamp\": \"2025-01-15T10:30:08Z\"}\n\nevent: step_completed\ndata: {\"step\": \"html_analysis\", \"progress\": 60, \"results\": {\"html_version\": \"HTML5\", \"title\": \"Example Domain\"}, \"timestamp\": \"2025-01-15T10:30:10Z\"}\n\nevent: progress\ndata: {\"step\": \"analyzing_links\", \"progress\": 75, \"message\": \"Analyzing links...\", \"timestamp\": \"2025-01-15T10:30:12Z\"}\n\nevent: step_completed\ndata: {\"step\": \"link_analysis\", \"progress\": 90, \"results\": {\"internal_count\": 15, \"external_count\": 8}, \"timestamp\": \"2025-01-15T10:30:14Z\"}\n\nevent: completed\ndata: {\"analysis_id\": \"550e8400-e29b-41d4-a716-446655440000\", \"status\": \"completed\", \"timestamp\": \"2025-01-15T10:30:15Z\"}\n"
                  },
                  "error_event": {
                    "summary": "SSE error event",
                    "value": "event: started\ndata: {\"analysis_id\": \"550e8400-e29b-41d4-a716-446655440001\", \"status\": \"started\", \"timestamp\": \"2025-01-15T10:30:00Z\"}\n\nevent: progress\ndata: {\"step\": \"fetching_page\", \"progress\": 25, \"message\": \"Fetching page content...\", \"timestamp\": \"2025-01-15T10:30:05Z\"}\n\nevent: error\ndata: {\"analysis_id\": \"550e8400-e29b-41d4-a716-446655440001\", \"status\": \"failed\", \"error\": \"page_unreachable\", \"message\": \"Connection timeout\", \"timestamp\": \"2025-01-15T10:30:30Z\"}\n"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "PasetoAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "PASETO",
        "description": "PASETO (Platform Authentication Security Token Exchange and Operations) v4 token.\nA secure token format with issuer validation and enhanced security features.\nFormat: Bearer v4.local.{payload} or Bearer v4.public.{payload}.{signature}\n"
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Basic HTTP authentication for administrative endpoints"
      }
    },
...
  }
}